completed 6step test
The results for the 6 step examination and its comparison the the 12 and 24 step results suggest a clear direction to take moving forward. After full training with the same LSTM neural network architecture, the smaller the step the better the results. makes sense as there is less noise of the 1-2 past hour range. Going to collect more data on sunday, looking to double data-set size. after doing that, will test for 6 and 12 step on same LSTM NN architecture, see if results are more in the same direction and then go from there.