data collection and more prog (rd)
currently running long test on Dr1, realizing I had class drop code wrong for when Dr1 was selected, meaning that it wasnt as hard as I thought to train a model on. using huge model with 70 second epochs to see what happens. The iterative process of training with data + model prediction is working out better than expected. May be able to have far more accurate model here soon. actually it might already be the v11 model for Dr3 just have to test it.